+++ migrations/token/metadata/TokenMetadataResolverRandomizedReveal@3.0_deploy.js
@@ -1,12 +1,13 @@
-const {getContractAddressOrZero} = require('../../../src/helpers/templates');
+const {ethers} = require('hardhat');
+const {getContractAddressOrFallback} = require('../../../src/helpers/templates');
 const Contract_deploy = require('../../../src/templates/Contract/deploy');
 
 module.exports = Contract_deploy('TokenMetadataResolverRandomizedReveal@3.0', {
   contract: 'TokenMetadataResolverRandomizedReveal',
   importPath: 'node_modules/@animoca/ethereum-contracts-3.0/artifacts',
   args: [
-    {name: 'Chainlink_LinkToken', value: getContractAddressOrZero('Chainlink_LinkToken')},
-    {name: 'Chainlink_VRFV2Wrapper', value: getContractAddressOrZero('Chainlink_VRFV2Wrapper')},
+    {name: 'Chainlink_LinkToken', value: getContractAddressOrFallback('Chainlink_LinkToken', ethers.constants.AddressZero)},
+    {name: 'Chainlink_VRFV2Wrapper', value: getContractAddressOrFallback('Chainlink_VRFV2Wrapper', ethers.constants.AddressZero)},
   ],
   deterministicDeployment: true,
 });